+++ src/time_split_app/widgets/parameters/_splitter_kwargs.py
@@ -14,6 +14,17 @@ class SplitterKwargsWidget:
     after_widget: SpanWidget = field(default_factory=lambda: SpanWidget("after"))
     limits_widget: ExpandLimitsWidget = field(default_factory=ExpandLimitsWidget)
 
+    def __post_init__(self) -> None:
+        before_widget = self.before_widget
+        if before_widget.span != "before":
+            msg = f"Bad {before_widget=}; should have span='before'."
+            raise ValueError(msg)
+
+        after_widget = self.after_widget
+        if after_widget.span != "after":
+            msg = f"Bad {after_widget=}; should have span='after'."
+            raise ValueError()
+
     def select_params(self) -> DatetimeIndexSplitterKwargs:
         schedule, filters = self.schedule_widget.get_schedule()
         before, after = select_spans(self.before_widget, after=self.after_widget)